refactor: Separation of Localisable Data Binding error text attributes
