+++ GovUkDesignSystem/Attributes/DataBinding/GovUkDataBindingErrorTextAttribute.cs
@@ -10,21 +10,5 @@ namespace GovUkDesignSystem.Attributes.DataBinding
     [AttributeUsage(AttributeTargets.Property)]
     public abstract class GovUkDataBindingErrorTextAttribute : Attribute
     {
-        private ResourceManager _resourceManager;
-
-        protected Type ResourceType { get; set; }
-
-        protected string ResourceName { get; set; }
-
-        protected string GetResourceValue(string resourceKey)
-        {
-            var resourceAssembly = ResourceType?.Assembly;
-            if (resourceAssembly != null && _resourceManager == null)
-            {
-                _resourceManager = new ResourceManager(ResourceName, resourceAssembly);
-            }
-
-            return _resourceManager?.GetString(resourceKey, CultureInfo.CurrentCulture) ?? resourceKey;
-        }
     }
 }
