+++ GovUkDesignSystem/Attributes/DataBinding/GovUkDataBindingMandatoryDecimalErrorTextAttribute.cs
@@ -4,28 +4,18 @@ namespace GovUkDesignSystem.Attributes.DataBinding
 {
     public class GovUkDataBindingMandatoryDecimalErrorTextAttribute : GovUkDataBindingErrorTextAttribute
     {
-        public GovUkDataBindingMandatoryDecimalErrorTextAttribute(string errorMessageIfMissing, string nameAtStartOfSentence = "", string mustBeNumberErrorMessage = "", string resourceName = "", Type resourceType = null)
+        public GovUkDataBindingMandatoryDecimalErrorTextAttribute(string errorMessageIfMissing, string nameAtStartOfSentence)
         {
-            if(string.IsNullOrEmpty(nameAtStartOfSentence) && (string.IsNullOrEmpty(mustBeNumberErrorMessage)))
+            if (string.IsNullOrEmpty(nameAtStartOfSentence))
             {
-                throw new ArgumentNullException("nameAtStartOfSentence cannot be null or empty unless all error messages are overidden");
+                throw new ArgumentNullException("nameAtStartOfSentence cannot be null or empty");
             }
-            
-            if (resourceType == null ^ string.IsNullOrEmpty(resourceName))
-            {
-                throw new ArgumentNullException("resourceName or resourceType cannot be null or empty while the other is not null or empty");
-            }
-            
             if(string.IsNullOrEmpty(errorMessageIfMissing))
             {
                 throw new ArgumentNullException("errorMessageIfMissing cannot be null or empty");
             }
-            
             NameAtStartOfSentence = nameAtStartOfSentence;
             ErrorMessageIfMissing = errorMessageIfMissing;
-            MustBeNumberErrorMessage = mustBeNumberErrorMessage;
-            ResourceType = resourceType;
-            ResourceName = resourceName;
         }
 
         /// <summary>
@@ -39,23 +29,6 @@ namespace GovUkDesignSystem.Attributes.DataBinding
         /// A complete sentence of the form: ‘Enter [whatever it is]’.
         /// <br/>For example, ‘Enter your first name’.
         /// </summary>
-        private string _errorMessageIfMissing;
-        public string ErrorMessageIfMissing
-        {
-            get => GetResourceValue(_errorMessageIfMissing);
-            private set => _errorMessageIfMissing = value;
-        }
-        
-        /// <summary>
-        /// An override for the error message that is displayed if the value entered is not a number.
-        /// A complete sentence of the form: ‘[Whatever it is] must be a number’
-        /// <br/>e.g. "Median age must be a number"
-        /// </summary>
-        private string _mustBeNumberErrorMessage;
-        public string MustBeNumberErrorMessage
-        {
-            get => GetResourceValue(_mustBeNumberErrorMessage);
-            private set => _mustBeNumberErrorMessage = value;
-        }
+        public string ErrorMessageIfMissing { get; private set; }
     }
 }
