+++ GovUkDesignSystem/Attributes/DataBinding/GovUkDataBindingMandatoryIntErrorTextAttribute.cs
@@ -1,59 +1,23 @@
 ﻿using System;
 
-
 namespace GovUkDesignSystem.Attributes.DataBinding
 {
     public class GovUkDataBindingMandatoryIntErrorTextAttribute : GovUkDataBindingErrorTextAttribute
     {
-        public GovUkDataBindingMandatoryIntErrorTextAttribute(string errorMessageIfMissing, string nameAtStartOfSentence = "", Type resourceType = null, string resourceName = "", string isWholeNumberErrorMessage = "", string mustBeNumberErrorMessage = "")
+        public GovUkDataBindingMandatoryIntErrorTextAttribute(string errorMessageIfMissing, string nameAtStartOfSentence)
         {
-            if(string.IsNullOrEmpty(nameAtStartOfSentence) && (string.IsNullOrEmpty(isWholeNumberErrorMessage) || string.IsNullOrEmpty(mustBeNumberErrorMessage)))
+            if(string.IsNullOrEmpty(nameAtStartOfSentence))
             {
-                throw new ArgumentNullException("nameAtStartOfSentence cannot be null or empty unless all error messages are overidden");
+                throw new ArgumentNullException("nameAtStartOfSentence cannot be null or empty");
             }
-
-            if (resourceType == null ^ string.IsNullOrEmpty(resourceName))
-            {
-                throw new ArgumentNullException("resourceName or resourceType cannot be null or empty while the other is not null or empty");
-            }
-            
             if(string.IsNullOrEmpty(errorMessageIfMissing))
             {
                 throw new ArgumentNullException("errorMessageIfMissing cannot be null or empty");
             }
-            
             NameAtStartOfSentence = nameAtStartOfSentence;
             ErrorMessageIfMissing = errorMessageIfMissing;
-            MustBeNumberErrorMessage = mustBeNumberErrorMessage;
-            IsWholeNumberErrorMessage = isWholeNumberErrorMessage;
-            ResourceType = resourceType;
-            ResourceName = resourceName;
         }
 
-        /// <summary>
-        /// An override for the error message that is displayed if the value entered is not a whole number.
-        /// A complete sentence of the form: ‘[Whatever it is] must be a whole number’
-        /// <br/>e.g. "Median age must be a whole number"
-        /// </summary>
-        private string _isWholeNumberErrorMessage;
-        public string IsWholeNumberErrorMessage
-        {
-            get => GetResourceValue(_isWholeNumberErrorMessage);
-            private set => _isWholeNumberErrorMessage = value;
-        }
-        
-        /// <summary>
-        /// An override for the error message that is displayed if the value entered is not a number.
-        /// A complete sentence of the form: ‘[Whatever it is] must be a number’
-        /// <br/>e.g. "Median age must be a number"
-        /// </summary>
-        private string _mustBeNumberErrorMessage;
-        public string MustBeNumberErrorMessage
-        {
-            get => GetResourceValue(_mustBeNumberErrorMessage);
-            private set => _mustBeNumberErrorMessage = value;
-        }
-        
         /// <summary>
         /// The name as it would appear at the start of a sentence
         /// <br/>e.g. "[Full name] must be 2 characters or more"
@@ -65,11 +29,6 @@ namespace GovUkDesignSystem.Attributes.DataBinding
         /// A complete sentence of the form: ‘Enter [whatever it is]’.
         /// <br/>For example, ‘Enter your first name’.
         /// </summary>
-        private string _errorMessageIfMissing;
-        public string ErrorMessageIfMissing
-        {
-            get => GetResourceValue(_errorMessageIfMissing);
-            private set => _errorMessageIfMissing = value;
-        }
+        public string ErrorMessageIfMissing { get; private set; }
     }
 }
